Establish tls

+++ routes/leaderRouter.js
@@ -5,6 +5,7 @@ const leaderRouter = express.Router();
 leaderRouter.use(bodyParser.json());
 
 const Leaders = require("../models/leaders");
+let authenticate = require("../authenticate");
 
 leaderRouter
   .route("/")
@@ -22,34 +23,46 @@ leaderRouter
         next(err);
       });
   })
-  .post((req, res, next) => {
-    Leaders.create(req.body)
-      .then(
-        (leader) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(leader);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => next(err));
-  })
-  .put((req, res, next) => {
-    res.statusCode = 403;
-    res.end("put not supported for /leaders");
-  })
-  .delete((req, res, next) => {
-    Leaders.remove({})
-      .then(
-        (resp) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(resp);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => next(err));
-  });
+  .post(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      Leaders.create(req.body)
+        .then(
+          (leader) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(leader);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => next(err));
+    }
+  )
+  .put(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      res.statusCode = 403;
+      res.end("put not supported for /leaders");
+    }
+  )
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      Leaders.remove({})
+        .then(
+          (resp) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(resp);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => next(err));
+    }
+  );
 
 leaderRouter
   .route("/:leaderId")
@@ -65,37 +78,51 @@ leaderRouter
       )
       .catch((err) => next(err));
   })
-  .post((req, res, next) => {
-    res.statusCode = 403;
-    res.end("POST operation not supported on /leaders/" + req.params.leaderId);
-  })
-  .put((req, res, next) => {
-    Leaders.findByIdAndUpdate(
-      req.params.leaderId,
-      { $set: req.body },
-      { new: true }
-    )
-      .then(
-        (leader) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(leader);
-        },
-        (err) => next(err)
+  .post(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      res.statusCode = 403;
+      res.end(
+        "POST operation not supported on /leaders/" + req.params.leaderId
+      );
+    }
+  )
+  .put(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      Leaders.findByIdAndUpdate(
+        req.params.leaderId,
+        { $set: req.body },
+        { new: true }
       )
-      .catch((err) => next(err));
-  })
-  .delete((req, res, next) => {
-    Leaders.findByIdAndRemove(req.params.leaderId)
-      .then(
-        (resp) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(resp);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => next(err));
-  });
+        .then(
+          (leader) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(leader);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => next(err));
+    }
+  )
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      Leaders.findByIdAndRemove(req.params.leaderId)
+        .then(
+          (resp) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(resp);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => next(err));
+    }
+  );
 
 module.exports = leaderRouter;
