+++ routes/promoRouter.js
@@ -5,6 +5,7 @@ const promoRouter = express.Router();
 promoRouter.use(bodyParser.json());
 
 const Promotions = require("../models/promotions");
+let authenticate = require("../authenticate");
 
 promoRouter
   .route("/")
@@ -22,34 +23,46 @@ promoRouter
         next(err);
       });
   })
-  .post((req, res, next) => {
-    Promotions.create(req.body)
-      .then(
-        (promo) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(promo);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => next(err));
-  })
-  .put((req, res, next) => {
-    res.statusCode = 403;
-    res.end("put not supported for /promotions");
-  })
-  .delete((req, res, next) => {
-    Promotions.remove({})
-      .then(
-        (resp) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(resp);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => next(err));
-  });
+  .post(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      Promotions.create(req.body)
+        .then(
+          (promo) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(promo);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => next(err));
+    }
+  )
+  .put(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      res.statusCode = 403;
+      res.end("put not supported for /promotions");
+    }
+  )
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      Promotions.remove({})
+        .then(
+          (resp) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(resp);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => next(err));
+    }
+  );
 
 promoRouter
   .route("/:promoId")
@@ -65,39 +78,51 @@ promoRouter
       )
       .catch((err) => next(err));
   })
-  .post((req, res, next) => {
-    res.statusCode = 403;
-    res.end(
-      "POST operation not supported on /promotions/" + req.params.promoId
-    );
-  })
-  .put((req, res, next) => {
-    Promotions.findByIdAndUpdate(
-      req.params.promoId,
-      { $set: req.body },
-      { new: true }
-    )
-      .then(
-        (promo) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(promo);
-        },
-        (err) => next(err)
+  .post(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      res.statusCode = 403;
+      res.end(
+        "POST operation not supported on /promotions/" + req.params.promoId
+      );
+    }
+  )
+  .put(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      Promotions.findByIdAndUpdate(
+        req.params.promoId,
+        { $set: req.body },
+        { new: true }
       )
-      .catch((err) => next(err));
-  })
-  .delete((req, res, next) => {
-    Promotions.findByIdAndRemove(req.params.promoId)
-      .then(
-        (resp) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(resp);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => next(err));
-  });
+        .then(
+          (promo) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(promo);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => next(err));
+    }
+  )
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin(),
+    (req, res, next) => {
+      Promotions.findByIdAndRemove(req.params.promoId)
+        .then(
+          (resp) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(resp);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => next(err));
+    }
+  );
 
 module.exports = promoRouter;
